+++ src/main/java/org/fundacionjala/sfdc/driver/DriverFactory.java
@@ -1,0 +1,40 @@
+package org.fundacionjala.sfdc.driver;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebDriverException;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+/**
+ * Created by abelb on 8/8/2017.
+ */
+public final class DriverFactory {
+
+    /**
+     * Private Constructor.
+     */
+    private DriverFactory() {
+
+    }
+
+    /**
+     * Return current Driver.
+     * @param driverType Enum from DriverType.
+     * @return Current WebDriver.
+     */
+    public static WebDriver getDriverManager(final DriverType driverType) {
+        WebDriver driver;
+
+        switch (driverType) {
+            case CHROME:
+                driver = new ChromeDriver();
+                break;
+            case FIREFOX:
+                driver = new FirefoxDriver();
+                break;
+            default:
+                throw new WebDriverException("No driver provided");
+        }
+        return driver;
+    }
+}
