+++ src/main/java/org/fundacionjala/sfdc/pages/login/Login.java
@@ -1,7 +1,7 @@
 package org.fundacionjala.sfdc.pages.login;
 
 import org.fundacionjala.sfdc.CommonActions;
-import org.fundacionjala.sfdc.DriverManager;
+import org.fundacionjala.sfdc.driver.DriverManager;
 import org.fundacionjala.sfdc.pages.Home;
 import org.fundacionjala.sfdc.pages.Profile;
 import org.fundacionjala.sfdc.pages.base.BasePage;
